// src/app/model/clientSubscription.ts
export class ClientSubscription {
    subscriptionId: number;
    planId: number;
    clientId: number;
    purchaseId: number;
    clientName: string;
    planName: string;
    purchaseDatetime: string;
    startDate: string;
    validity: number;
    payment_mode: string;
    transactionNo: string;
    endDate: string;
    amount: number;
    price: number;
    updDatetime: number;
    updUserId: number;
    status: number;

    constructor() {
        
    }


}